:goal_net: More specific error handling
